Implements MountFS
