+++ packages/yarnpkg-fslib/sources/ZipOpenFS.ts
@@ -1,32 +1,16 @@
-import {Libzip}                                                                                                                                      from '@yarnpkg/libzip';
-import {BigIntStats, constants, Stats}                                                                                                               from 'fs';
+import {Libzip}                                   from '@yarnpkg/libzip';
 
-import {WatchOptions, WatchCallback, Watcher, StatOptions, StatSyncOptions}                                                                          from './FakeFS';
-import {FakeFS, MkdirOptions, RmdirOptions, WriteFileOptions, OpendirOptions}                                                                        from './FakeFS';
-import {Dirent, SymlinkType}                                                                                                                         from './FakeFS';
-import {CreateReadStreamOptions, CreateWriteStreamOptions, BasePortableFakeFS, ExtractHintOptions, WatchFileOptions, WatchFileCallback, StatWatcher} from './FakeFS';
-import {NodeFS}                                                                                                                                      from './NodeFS';
-import {ZipFS}                                                                                                                                       from './ZipFS';
-import {watchFile, unwatchFile, unwatchAllFiles}                                                                                                     from './algorithms/watchFile';
-import * as errors                                                                                                                                   from './errors';
-import {Filename, FSPath, npath, PortablePath, ppath}                                                                                                from './path';
-
-// Only file descriptors prefixed by those values will be forwarded to the ZipFS
-// instances. Note that the highest ZIP_MAGIC bit MUST NOT be set, otherwise the
-// resulting fd becomes a negative integer, which isn't supposed to happen per
-// the unix rules (caused problems w/ Go).
-//
-// Those values must be synced with packages/yarnpkg-pnp/sources/esm-loader/fspatch.ts
-//
-const ZIP_MASK  = 0xff000000;
-const ZIP_MAGIC = 0x2a000000;
+import {FakeFS}                                   from './FakeFS';
+import {GetMountPointFn, MountFS, MountFSOptions} from './MountFS';
+import {ZipFS}                                    from './ZipFS';
+import {PortablePath, ppath}                      from './path';
 
 /**
  * Extracts the archive part (ending in the first instance of `extension`) from a path.
  *
  * The indexOf-based implementation is ~3.7x faster than a RegExp-based implementation.
  */
-export const getArchivePart = (path: string, extension: string) => {
+function getArchivePart(path: string, extension: string) {
   let idx = path.indexOf(extension);
   if (idx <= 0)
     return null;
@@ -49,20 +33,12 @@ export const getArchivePart = (path: string, extension: string) => {
     return null;
 
   return path.slice(0, nextCharIdx) as PortablePath;
-};
+}
 
-export type ZipOpenFSOptions = {
-  baseFs?: FakeFS<PortablePath>;
-  filter?: RegExp | null;
+export type ZipOpenFSOptions = MountFSOptions<ZipFS> & {
   libzip: Libzip | (() => Libzip);
-  maxOpenFiles?: number;
   readOnlyArchives?: boolean;
-  useCache?: boolean;
-  /**
-   * Maximum age in ms.
-   * ZipFS instances are pruned from the cache if they aren't accessed within this amount of time.
-   */
-  maxAge?: number;
+
   /**
    * Which file extensions will be interpreted as zip files. Useful for supporting other formats
    * packaged as zips, such as .docx.
@@ -72,7 +48,7 @@ export type ZipOpenFSOptions = {
   fileExtensions?: Array<string> | null;
 };
 
-export class ZipOpenFS extends BasePortableFakeFS {
+export class ZipOpenFS extends MountFS<ZipFS> {
   static async openPromise<T>(fn: (zipOpenFs: ZipOpenFS) => Promise<T>, opts: ZipOpenFSOptions): Promise<T> {
     const zipOpenFs = new ZipOpenFS(opts);
 
@@ -83,1045 +59,65 @@ export class ZipOpenFS extends BasePortableFakeFS {
     }
   }
 
-  private libzipFactory!: () => Libzip;
-  private libzipInstance?: Libzip;
+  constructor(opts: ZipOpenFSOptions) {
+    let libzipInstance: Libzip | undefined;
 
-  private get libzip(): Libzip {
-    if (typeof this.libzipInstance === `undefined`)
-      this.libzipInstance = this.libzipFactory();
+    const libzipFactory: () => Libzip = typeof opts.libzip !== `function`
+      ? () => opts.libzip as Exclude<(typeof opts)[`libzip`], Function>
+      : opts.libzip;
 
-    return this.libzipInstance;
-  }
+    const getLibzip = () => {
+      if (typeof libzipInstance === `undefined`)
+        libzipInstance = libzipFactory();
 
-  private readonly baseFs: FakeFS<PortablePath>;
-
-  private readonly zipInstances: Map<string, {zipFs: ZipFS, expiresAt: number, refCount: number}> | null;
-
-  private readonly fdMap: Map<number, [ZipFS, number]> = new Map();
-  private nextFd = 3;
-
-  private readonly filter: RegExp | null;
-  private readonly maxOpenFiles: number;
-  private readonly maxAge: number;
-  private readonly readOnlyArchives: boolean;
-  private readonly fileExtensions: Array<string> | null;
-
-  private isZip: Set<PortablePath> = new Set();
-  private notZip: Set<PortablePath> = new Set();
-  private realPaths: Map<PortablePath, PortablePath> = new Map();
-
-  constructor({libzip, baseFs = new NodeFS(), filter = null, maxOpenFiles = Infinity, readOnlyArchives = false, useCache = true, maxAge = 5000, fileExtensions = null}: ZipOpenFSOptions) {
-    super();
-
-    this.libzipFactory = typeof libzip !== `function`
-      ? () => libzip
-      : libzip;
-
-    this.baseFs = baseFs;
-
-    this.zipInstances = useCache ? new Map() : null;
-
-    this.filter = filter;
-    this.maxOpenFiles = maxOpenFiles;
-    this.readOnlyArchives = readOnlyArchives;
-    this.maxAge = maxAge;
-    this.fileExtensions = fileExtensions;
-  }
-
-  getExtractHint(hints: ExtractHintOptions) {
-    return this.baseFs.getExtractHint(hints);
-  }
-
-  getRealPath() {
-    return this.baseFs.getRealPath();
-  }
-
-  saveAndClose() {
-    unwatchAllFiles(this);
-
-    if (this.zipInstances) {
-      for (const [path, {zipFs}] of this.zipInstances.entries()) {
-        zipFs.saveAndClose();
-        this.zipInstances.delete(path);
-      }
-    }
-  }
-
-  discardAndClose() {
-    unwatchAllFiles(this);
-
-    if (this.zipInstances) {
-      for (const [path, {zipFs}] of this.zipInstances.entries()) {
-        zipFs.discardAndClose();
-        this.zipInstances.delete(path);
-      }
-    }
-  }
-
-  resolve(p: PortablePath) {
-    return this.baseFs.resolve(p);
-  }
-
-  private remapFd(zipFs: ZipFS, fd: number) {
-    const remappedFd = this.nextFd++ | ZIP_MAGIC;
-    this.fdMap.set(remappedFd, [zipFs, fd]);
-    return remappedFd;
-  }
-
-  async openPromise(p: PortablePath, flags: string, mode?: number) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.openPromise(p, flags, mode);
-    }, async (zipFs, {subPath}) => {
-      return this.remapFd(zipFs, await zipFs.openPromise(subPath, flags, mode));
-    });
-  }
-
-  openSync(p: PortablePath, flags: string, mode?: number) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.openSync(p, flags, mode);
-    }, (zipFs, {subPath}) => {
-      return this.remapFd(zipFs, zipFs.openSync(subPath, flags, mode));
-    });
-  }
-
-  async opendirPromise(p: PortablePath, opts?: OpendirOptions) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.opendirPromise(p, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.opendirPromise(subPath, opts);
-    }, {
-      requireSubpath: false,
-    });
-  }
-
-  opendirSync(p: PortablePath, opts?: OpendirOptions) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.opendirSync(p, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.opendirSync(subPath, opts);
-    }, {
-      requireSubpath: false,
-    });
-  }
-
-  async readPromise(fd: number, buffer: Buffer, offset: number, length: number, position: number) {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return await this.baseFs.readPromise(fd, buffer, offset, length, position);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`read`);
-
-    const [zipFs, realFd] = entry;
-    return await zipFs.readPromise(realFd, buffer, offset, length, position);
-  }
-
-  readSync(fd: number, buffer: Buffer, offset: number, length: number, position: number) {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.readSync(fd, buffer, offset, length, position);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`readSync`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.readSync(realFd, buffer, offset, length, position);
-  }
-
-  writePromise(fd: number, buffer: Buffer, offset?: number, length?: number, position?: number): Promise<number>;
-  writePromise(fd: number, buffer: string, position?: number): Promise<number>;
-  async writePromise(fd: number, buffer: Buffer | string, offset?: number, length?: number, position?: number): Promise<number> {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC) {
-      if (typeof buffer === `string`) {
-        return await this.baseFs.writePromise(fd, buffer, offset);
-      } else {
-        return await this.baseFs.writePromise(fd, buffer, offset, length, position);
-      }
-    }
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`write`);
-
-    const [zipFs, realFd] = entry;
-
-    if (typeof buffer === `string`) {
-      return await zipFs.writePromise(realFd, buffer, offset);
-    } else {
-      return await zipFs.writePromise(realFd, buffer, offset, length, position);
-    }
-  }
-
-  writeSync(fd: number, buffer: Buffer, offset?: number, length?: number, position?: number): number;
-  writeSync(fd: number, buffer: string, position?: number): number;
-  writeSync(fd: number, buffer: Buffer | string, offset?: number, length?: number, position?: number): number {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC) {
-      if (typeof buffer === `string`) {
-        return this.baseFs.writeSync(fd, buffer, offset);
-      } else {
-        return this.baseFs.writeSync(fd, buffer, offset, length, position);
-      }
-    }
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`writeSync`);
-
-    const [zipFs, realFd] = entry;
-
-    if (typeof buffer === `string`) {
-      return zipFs.writeSync(realFd, buffer, offset);
-    } else {
-      return zipFs.writeSync(realFd, buffer, offset, length, position);
-    }
-  }
-
-  async closePromise(fd: number) {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return await this.baseFs.closePromise(fd);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`close`);
-
-    this.fdMap.delete(fd);
-
-    const [zipFs, realFd] = entry;
-    return await zipFs.closePromise(realFd);
-  }
-
-  closeSync(fd: number) {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.closeSync(fd);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`closeSync`);
-
-    this.fdMap.delete(fd);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.closeSync(realFd);
-  }
-
-  createReadStream(p: PortablePath | null, opts?: CreateReadStreamOptions) {
-    if (p === null)
-      return this.baseFs.createReadStream(p, opts);
-
-    return this.makeCallSync(p, () => {
-      return this.baseFs.createReadStream(p, opts);
-    }, (zipFs, {archivePath, subPath}) => {
-      const stream = zipFs.createReadStream(subPath, opts);
-      // This is a very hacky workaround. `ZipOpenFS` shouldn't have to work with `NativePath`s.
-      // Ref: https://github.com/yarnpkg/berry/pull/3774
-      // TODO: think of a better solution
-      stream.path = npath.fromPortablePath(this.pathUtils.join(archivePath, subPath));
-      return stream;
-    });
-  }
-
-  createWriteStream(p: PortablePath | null, opts?: CreateWriteStreamOptions) {
-    if (p === null)
-      return this.baseFs.createWriteStream(p, opts);
-
-    return this.makeCallSync(p, () => {
-      return this.baseFs.createWriteStream(p, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.createWriteStream(subPath, opts);
-    });
-  }
-
-  async realpathPromise(p: PortablePath) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.realpathPromise(p);
-    }, async (zipFs, {archivePath, subPath}) => {
-      let realArchivePath = this.realPaths.get(archivePath);
-      if (typeof realArchivePath === `undefined`) {
-        realArchivePath = await this.baseFs.realpathPromise(archivePath);
-        this.realPaths.set(archivePath, realArchivePath);
-      }
-
-      return this.pathUtils.join(realArchivePath, this.pathUtils.relative(PortablePath.root, await zipFs.realpathPromise(subPath)));
-    });
-  }
-
-  realpathSync(p: PortablePath) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.realpathSync(p);
-    }, (zipFs, {archivePath, subPath}) => {
-      let realArchivePath = this.realPaths.get(archivePath);
-      if (typeof realArchivePath === `undefined`) {
-        realArchivePath = this.baseFs.realpathSync(archivePath);
-        this.realPaths.set(archivePath, realArchivePath);
-      }
-
-      return this.pathUtils.join(realArchivePath, this.pathUtils.relative(PortablePath.root, zipFs.realpathSync(subPath)));
-    });
-  }
-
-  async existsPromise(p: PortablePath) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.existsPromise(p);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.existsPromise(subPath);
-    });
-  }
-
-  existsSync(p: PortablePath) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.existsSync(p);
-    }, (zipFs, {subPath}) => {
-      return zipFs.existsSync(subPath);
-    });
-  }
-
-  async accessPromise(p: PortablePath, mode?: number) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.accessPromise(p, mode);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.accessPromise(subPath, mode);
-    });
-  }
-
-  accessSync(p: PortablePath, mode?: number) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.accessSync(p, mode);
-    }, (zipFs, {subPath}) => {
-      return zipFs.accessSync(subPath, mode);
-    });
-  }
-
-  // https://github.com/DefinitelyTyped/DefinitelyTyped/blob/51d793492d4c2e372b01257668dcd3afc58d7352/types/node/v16/fs.d.ts#L1042-L1059
-  async statPromise(p: PortablePath): Promise<Stats>;
-  async statPromise(p: PortablePath, opts: (StatOptions & { bigint?: false | undefined }) | undefined): Promise<Stats>;
-  async statPromise(p: PortablePath, opts: StatOptions & { bigint: true }): Promise<BigIntStats>;
-  async statPromise(p: PortablePath, opts?: StatOptions): Promise<Stats | BigIntStats>;
-  async statPromise(p: PortablePath, opts?: StatOptions): Promise<Stats | BigIntStats> {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.statPromise(p, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.statPromise(subPath, opts);
-    });
-  }
-
-  // https://github.com/DefinitelyTyped/DefinitelyTyped/blob/51d793492d4c2e372b01257668dcd3afc58d7352/types/node/v16/fs.d.ts#L931-L967
-  statSync(p: PortablePath): Stats;
-  statSync(p: PortablePath, opts?: StatSyncOptions & {bigint?: false | undefined, throwIfNoEntry: false}): Stats | undefined;
-  statSync(p: PortablePath, opts: StatSyncOptions & {bigint: true, throwIfNoEntry: false}): BigIntStats | undefined;
-  statSync(p: PortablePath, opts?: StatSyncOptions & {bigint?: false | undefined}): Stats;
-  statSync(p: PortablePath, opts: StatSyncOptions & {bigint: true}): BigIntStats;
-  statSync(p: PortablePath, opts: StatSyncOptions & {bigint: boolean, throwIfNoEntry?: false | undefined}): Stats | BigIntStats;
-  statSync(p: PortablePath, opts?: StatSyncOptions): Stats | BigIntStats | undefined {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.statSync(p, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.statSync(subPath, opts);
-    });
-  }
-
-  async fstatPromise(fd: number): Promise<Stats>
-  async fstatPromise(fd: number, opts: {bigint: true}): Promise<BigIntStats>
-  async fstatPromise(fd: number, opts?: {bigint: boolean}): Promise<BigIntStats | Stats>
-  async fstatPromise(fd: number, opts?: { bigint: boolean }) {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.fstatPromise(fd, opts);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`fstat`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.fstatPromise(realFd, opts);
-  }
-
-  fstatSync(fd: number): Stats
-  fstatSync(fd: number, opts: {bigint: true}): BigIntStats
-  fstatSync(fd: number, opts?: {bigint: boolean}): BigIntStats | Stats
-  fstatSync(fd: number, opts?: { bigint: boolean }) {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.fstatSync(fd, opts);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`fstatSync`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.fstatSync(realFd, opts);
-  }
-
-  // https://github.com/DefinitelyTyped/DefinitelyTyped/blob/51d793492d4c2e372b01257668dcd3afc58d7352/types/node/v16/fs.d.ts#L1042-L1059
-  async lstatPromise(p: PortablePath): Promise<Stats>;
-  async lstatPromise(p: PortablePath, opts: (StatOptions & { bigint?: false | undefined }) | undefined): Promise<Stats>;
-  async lstatPromise(p: PortablePath, opts: StatOptions & { bigint: true }): Promise<BigIntStats>;
-  async lstatPromise(p: PortablePath, opts?: StatOptions): Promise<Stats | BigIntStats> {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.lstatPromise(p, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.lstatPromise(subPath, opts);
-    });
-  }
-
-  // https://github.com/DefinitelyTyped/DefinitelyTyped/blob/51d793492d4c2e372b01257668dcd3afc58d7352/types/node/v16/fs.d.ts#L931-L967
-  lstatSync(p: PortablePath): Stats;
-  lstatSync(p: PortablePath, opts?: StatSyncOptions & {bigint?: false | undefined, throwIfNoEntry: false}): Stats | undefined;
-  lstatSync(p: PortablePath, opts: StatSyncOptions & {bigint: true, throwIfNoEntry: false}): BigIntStats | undefined;
-  lstatSync(p: PortablePath, opts?: StatSyncOptions & {bigint?: false | undefined}): Stats;
-  lstatSync(p: PortablePath, opts: StatSyncOptions & {bigint: true}): BigIntStats;
-  lstatSync(p: PortablePath, opts: StatSyncOptions & { bigint: boolean, throwIfNoEntry?: false | undefined }): Stats | BigIntStats;
-  lstatSync(p: PortablePath, opts?: StatSyncOptions): Stats | BigIntStats | undefined {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.lstatSync(p, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.lstatSync(subPath, opts);
-    });
-  }
-
-  async fchmodPromise(fd: number, mask: number): Promise<void> {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.fchmodPromise(fd, mask);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`fchmod`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.fchmodPromise(realFd, mask);
-  }
-
-  fchmodSync(fd: number, mask: number): void {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.fchmodSync(fd, mask);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`fchmodSync`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.fchmodSync(realFd, mask);
-  }
-
-  async chmodPromise(p: PortablePath, mask: number) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.chmodPromise(p, mask);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.chmodPromise(subPath, mask);
-    });
-  }
-
-  chmodSync(p: PortablePath, mask: number) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.chmodSync(p, mask);
-    }, (zipFs, {subPath}) => {
-      return zipFs.chmodSync(subPath, mask);
-    });
-  }
-
-  async chownPromise(p: PortablePath, uid: number, gid: number) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.chownPromise(p, uid, gid);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.chownPromise(subPath, uid, gid);
-    });
-  }
-
-  chownSync(p: PortablePath, uid: number, gid: number) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.chownSync(p, uid, gid);
-    }, (zipFs, {subPath}) => {
-      return zipFs.chownSync(subPath, uid, gid);
-    });
-  }
-
-  async renamePromise(oldP: PortablePath, newP: PortablePath) {
-    return await this.makeCallPromise(oldP, async () => {
-      return await this.makeCallPromise(newP, async () => {
-        return await this.baseFs.renamePromise(oldP, newP);
-      }, async () => {
-        throw Object.assign(new Error(`EEXDEV: cross-device link not permitted`), {code: `EEXDEV`});
-      });
-    }, async (zipFsO, {subPath: subPathO}) => {
-      return await this.makeCallPromise(newP, async () => {
-        throw Object.assign(new Error(`EEXDEV: cross-device link not permitted`), {code: `EEXDEV`});
-      }, async (zipFsN, {subPath: subPathN}) => {
-        if (zipFsO !== zipFsN) {
-          throw Object.assign(new Error(`EEXDEV: cross-device link not permitted`), {code: `EEXDEV`});
-        } else {
-          return await zipFsO.renamePromise(subPathO, subPathN);
-        }
-      });
-    });
-  }
-
-  renameSync(oldP: PortablePath, newP: PortablePath) {
-    return this.makeCallSync(oldP, () => {
-      return this.makeCallSync(newP, () => {
-        return this.baseFs.renameSync(oldP, newP);
-      }, () => {
-        throw Object.assign(new Error(`EEXDEV: cross-device link not permitted`), {code: `EEXDEV`});
-      });
-    }, (zipFsO, {subPath: subPathO}) => {
-      return this.makeCallSync(newP, () => {
-        throw Object.assign(new Error(`EEXDEV: cross-device link not permitted`), {code: `EEXDEV`});
-      }, (zipFsN, {subPath: subPathN}) => {
-        if (zipFsO !== zipFsN) {
-          throw Object.assign(new Error(`EEXDEV: cross-device link not permitted`), {code: `EEXDEV`});
-        } else {
-          return zipFsO.renameSync(subPathO, subPathN);
-        }
-      });
-    });
-  }
-
-  async copyFilePromise(sourceP: PortablePath, destP: PortablePath, flags: number = 0) {
-    const fallback = async (sourceFs: FakeFS<PortablePath>, sourceP: PortablePath, destFs: FakeFS<PortablePath>, destP: PortablePath) => {
-      if ((flags & constants.COPYFILE_FICLONE_FORCE) !== 0)
-        throw Object.assign(new Error(`EXDEV: cross-device clone not permitted, copyfile '${sourceP}' -> ${destP}'`), {code: `EXDEV`});
-      if ((flags & constants.COPYFILE_EXCL) && await this.existsPromise(sourceP))
-        throw Object.assign(new Error(`EEXIST: file already exists, copyfile '${sourceP}' -> '${destP}'`), {code: `EEXIST`});
-
-      let content;
-      try {
-        content = await sourceFs.readFilePromise(sourceP);
-      } catch (error) {
-        throw Object.assign(new Error(`EINVAL: invalid argument, copyfile '${sourceP}' -> '${destP}'`), {code: `EINVAL`});
-      }
-
-      await destFs.writeFilePromise(destP, content);
+      return libzipInstance;
     };
 
-    return await this.makeCallPromise(sourceP, async () => {
-      return await this.makeCallPromise(destP, async () => {
-        return await this.baseFs.copyFilePromise(sourceP, destP, flags);
-      }, async (zipFsD, {subPath: subPathD}) => {
-        return await fallback(this.baseFs, sourceP, zipFsD, subPathD);
-      });
-    }, async (zipFsS, {subPath: subPathS}) => {
-      return await this.makeCallPromise(destP, async () => {
-        return await fallback(zipFsS, subPathS, this.baseFs, destP);
-      }, async (zipFsD, {subPath: subPathD}) => {
-        if (zipFsS !== zipFsD) {
-          return await fallback(zipFsS, subPathS, zipFsD, subPathD);
-        } else {
-          return await zipFsS.copyFilePromise(subPathS, subPathD, flags);
-        }
-      });
-    });
-  }
+    const fileExtensions = opts.fileExtensions;
+    const readOnlyArchives = opts.readOnlyArchives;
 
-  copyFileSync(sourceP: PortablePath, destP: PortablePath, flags: number = 0) {
-    const fallback = (sourceFs: FakeFS<PortablePath>, sourceP: PortablePath, destFs: FakeFS<PortablePath>, destP: PortablePath) => {
-      if ((flags & constants.COPYFILE_FICLONE_FORCE) !== 0)
-        throw Object.assign(new Error(`EXDEV: cross-device clone not permitted, copyfile '${sourceP}' -> ${destP}'`), {code: `EXDEV`});
-      if ((flags & constants.COPYFILE_EXCL) && this.existsSync(sourceP))
-        throw Object.assign(new Error(`EEXIST: file already exists, copyfile '${sourceP}' -> '${destP}'`), {code: `EEXIST`});
-
-      let content;
-      try {
-        content = sourceFs.readFileSync(sourceP);
-      } catch (error) {
-        throw Object.assign(new Error(`EINVAL: invalid argument, copyfile '${sourceP}' -> '${destP}'`), {code: `EINVAL`});
-      }
-
-      destFs.writeFileSync(destP, content);
-    };
-
-    return this.makeCallSync(sourceP, () => {
-      return this.makeCallSync(destP, () => {
-        return this.baseFs.copyFileSync(sourceP, destP, flags);
-      }, (zipFsD, {subPath: subPathD}) => {
-        return fallback(this.baseFs, sourceP, zipFsD, subPathD);
-      });
-    }, (zipFsS, {subPath: subPathS}) => {
-      return this.makeCallSync(destP, () => {
-        return fallback(zipFsS, subPathS, this.baseFs, destP);
-      }, (zipFsD, {subPath: subPathD}) => {
-        if (zipFsS !== zipFsD) {
-          return fallback(zipFsS, subPathS, zipFsD, subPathD);
-        } else {
-          return zipFsS.copyFileSync(subPathS, subPathD, flags);
-        }
-      });
-    });
-  }
-
-  async appendFilePromise(p: FSPath<PortablePath>, content: string | Uint8Array, opts?: WriteFileOptions) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.appendFilePromise(p, content, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.appendFilePromise(subPath, content, opts);
-    });
-  }
-
-  appendFileSync(p: FSPath<PortablePath>, content: string | Uint8Array, opts?: WriteFileOptions) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.appendFileSync(p, content, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.appendFileSync(subPath, content, opts);
-    });
-  }
-
-  async writeFilePromise(p: FSPath<PortablePath>, content: string | NodeJS.ArrayBufferView, opts?: WriteFileOptions) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.writeFilePromise(p, content, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.writeFilePromise(subPath, content, opts);
-    });
-  }
-
-  writeFileSync(p: FSPath<PortablePath>, content: string | NodeJS.ArrayBufferView, opts?: WriteFileOptions) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.writeFileSync(p, content, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.writeFileSync(subPath, content, opts);
-    });
-  }
-
-  async unlinkPromise(p: PortablePath) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.unlinkPromise(p);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.unlinkPromise(subPath);
-    });
-  }
-
-  unlinkSync(p: PortablePath) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.unlinkSync(p);
-    }, (zipFs, {subPath}) => {
-      return zipFs.unlinkSync(subPath);
-    });
-  }
-
-  async utimesPromise(p: PortablePath, atime: Date | string | number, mtime: Date | string | number) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.utimesPromise(p, atime, mtime);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.utimesPromise(subPath, atime, mtime);
-    });
-  }
-
-  utimesSync(p: PortablePath, atime: Date | string | number, mtime: Date | string | number) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.utimesSync(p, atime, mtime);
-    }, (zipFs, {subPath}) => {
-      return zipFs.utimesSync(subPath, atime, mtime);
-    });
-  }
-
-  async mkdirPromise(p: PortablePath, opts?: MkdirOptions) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.mkdirPromise(p, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.mkdirPromise(subPath, opts);
-    });
-  }
-
-  mkdirSync(p: PortablePath, opts?: MkdirOptions) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.mkdirSync(p, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.mkdirSync(subPath, opts);
-    });
-  }
-
-  async rmdirPromise(p: PortablePath, opts?: RmdirOptions) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.rmdirPromise(p, opts);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.rmdirPromise(subPath, opts);
-    });
-  }
-
-  rmdirSync(p: PortablePath, opts?: RmdirOptions) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.rmdirSync(p, opts);
-    }, (zipFs, {subPath}) => {
-      return zipFs.rmdirSync(subPath, opts);
-    });
-  }
-
-  async linkPromise(existingP: PortablePath, newP: PortablePath) {
-    return await this.makeCallPromise(newP, async () => {
-      return await this.baseFs.linkPromise(existingP, newP);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.linkPromise(existingP, subPath);
-    });
-  }
-
-  linkSync(existingP: PortablePath, newP: PortablePath) {
-    return this.makeCallSync(newP, () => {
-      return this.baseFs.linkSync(existingP, newP);
-    }, (zipFs, {subPath}) => {
-      return zipFs.linkSync(existingP, subPath);
-    });
-  }
-
-  async symlinkPromise(target: PortablePath, p: PortablePath, type?: SymlinkType) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.symlinkPromise(target, p, type);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.symlinkPromise(target, subPath);
-    });
-  }
-
-  symlinkSync(target: PortablePath, p: PortablePath, type?: SymlinkType) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.symlinkSync(target, p, type);
-    }, (zipFs, {subPath}) => {
-      return zipFs.symlinkSync(target, subPath);
-    });
-  }
-
-  readFilePromise(p: FSPath<PortablePath>, encoding?: null): Promise<Buffer>;
-  readFilePromise(p: FSPath<PortablePath>, encoding: BufferEncoding): Promise<string>;
-  readFilePromise(p: FSPath<PortablePath>, encoding?: BufferEncoding | null): Promise<Buffer | string>;
-  async readFilePromise(p: FSPath<PortablePath>, encoding?: BufferEncoding | null) {
-    return this.makeCallPromise(p, async () => {
-      return await this.baseFs.readFilePromise(p, encoding);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.readFilePromise(subPath, encoding);
-    });
-  }
-
-  readFileSync(p: FSPath<PortablePath>, encoding?: null): Buffer;
-  readFileSync(p: FSPath<PortablePath>, encoding: BufferEncoding): string;
-  readFileSync(p: FSPath<PortablePath>, encoding?: BufferEncoding | null): Buffer | string;
-  readFileSync(p: FSPath<PortablePath>, encoding?: BufferEncoding | null) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.readFileSync(p, encoding);
-    }, (zipFs, {subPath}) => {
-      return zipFs.readFileSync(subPath, encoding);
-    });
-  }
-
-  async readdirPromise(p: PortablePath): Promise<Array<Filename>>;
-  async readdirPromise(p: PortablePath, opts: {withFileTypes: false} | null): Promise<Array<Filename>>;
-  async readdirPromise(p: PortablePath, opts: {withFileTypes: true}): Promise<Array<Dirent>>;
-  async readdirPromise(p: PortablePath, opts: {withFileTypes: boolean}): Promise<Array<Filename> | Array<Dirent>>;
-  async readdirPromise(p: PortablePath, opts?: {withFileTypes?: boolean} | null): Promise<Array<string> | Array<Dirent>> {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.readdirPromise(p, opts as any);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.readdirPromise(subPath, opts as any);
-    }, {
-      requireSubpath: false,
-    });
-  }
-
-  readdirSync(p: PortablePath): Array<Filename>;
-  readdirSync(p: PortablePath, opts: {withFileTypes: false} | null): Array<Filename>;
-  readdirSync(p: PortablePath, opts: {withFileTypes: true}): Array<Dirent>;
-  readdirSync(p: PortablePath, opts: {withFileTypes: boolean}): Array<Filename> | Array<Dirent>;
-  readdirSync(p: PortablePath, opts?: {withFileTypes?: boolean} | null): Array<string> | Array<Dirent> {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.readdirSync(p, opts as any);
-    }, (zipFs, {subPath}) => {
-      return zipFs.readdirSync(subPath, opts as any);
-    }, {
-      requireSubpath: false,
-    });
-  }
-
-  async readlinkPromise(p: PortablePath) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.readlinkPromise(p);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.readlinkPromise(subPath);
-    });
-  }
-
-  readlinkSync(p: PortablePath) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.readlinkSync(p);
-    }, (zipFs, {subPath}) => {
-      return zipFs.readlinkSync(subPath);
-    });
-  }
-
-  async truncatePromise(p: PortablePath, len?: number) {
-    return await this.makeCallPromise(p, async () => {
-      return await this.baseFs.truncatePromise(p, len);
-    }, async (zipFs, {subPath}) => {
-      return await zipFs.truncatePromise(subPath, len);
-    });
-  }
-
-  truncateSync(p: PortablePath, len?: number) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.truncateSync(p, len);
-    }, (zipFs, {subPath}) => {
-      return zipFs.truncateSync(subPath, len);
-    });
-  }
-
-  async ftruncatePromise(fd: number, len?: number): Promise<void> {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.ftruncatePromise(fd, len);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`ftruncate`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.ftruncatePromise(realFd, len);
-  }
-
-  ftruncateSync(fd: number, len?: number): void {
-    if ((fd & ZIP_MASK) !== ZIP_MAGIC)
-      return this.baseFs.ftruncateSync(fd, len);
-
-    const entry = this.fdMap.get(fd);
-    if (typeof entry === `undefined`)
-      throw errors.EBADF(`ftruncateSync`);
-
-    const [zipFs, realFd] = entry;
-    return zipFs.ftruncateSync(realFd, len);
-  }
-
-  watch(p: PortablePath, cb?: WatchCallback): Watcher;
-  watch(p: PortablePath, opts: WatchOptions, cb?: WatchCallback): Watcher;
-  watch(p: PortablePath, a?: WatchOptions | WatchCallback, b?: WatchCallback) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.watch(
-        p,
-        // @ts-expect-error
-        a,
-        b,
-      );
-    }, (zipFs, {subPath}) => {
-      return zipFs.watch(
-        subPath,
-        // @ts-expect-error
-        a,
-        b,
-      );
-    });
-  }
-
-  watchFile(p: PortablePath, cb: WatchFileCallback): StatWatcher;
-  watchFile(p: PortablePath, opts: WatchFileOptions, cb: WatchFileCallback): StatWatcher;
-  watchFile(p: PortablePath, a: WatchFileOptions | WatchFileCallback, b?: WatchFileCallback) {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.watchFile(
-        p,
-        // @ts-expect-error
-        a,
-        b,
-      );
-    }, () => {
-      return watchFile(this, p, a, b);
-    });
-  }
-
-  unwatchFile(p: PortablePath, cb?: WatchFileCallback): void {
-    return this.makeCallSync(p, () => {
-      return this.baseFs.unwatchFile(p, cb);
-    }, () => {
-      return unwatchFile(this, p, cb);
-    });
-  }
-
-  private async makeCallPromise<T>(p: FSPath<PortablePath>, discard: () => Promise<T>, accept: (zipFS: ZipFS, zipInfo: {archivePath: PortablePath, subPath: PortablePath}) => Promise<T>, {requireSubpath = true}: {requireSubpath?: boolean} = {}): Promise<T> {
-    if (typeof p !== `string`)
-      return await discard();
-
-    const normalizedP = this.resolve(p);
-
-    const zipInfo = this.findZip(normalizedP);
-    if (!zipInfo)
-      return await discard();
-
-    if (requireSubpath && zipInfo.subPath === `/`)
-      return await discard();
-
-    return await this.getZipPromise(zipInfo.archivePath, async zipFs => await accept(zipFs, zipInfo));
-  }
-
-  private makeCallSync<T>(p: FSPath<PortablePath>, discard: () => T, accept: (zipFS: ZipFS, zipInfo: {archivePath: PortablePath, subPath: PortablePath}) => T, {requireSubpath = true}: {requireSubpath?: boolean} = {}): T {
-    if (typeof p !== `string`)
-      return discard();
-
-    const normalizedP = this.resolve(p);
-
-    const zipInfo = this.findZip(normalizedP);
-    if (!zipInfo)
-      return discard();
-
-    if (requireSubpath && zipInfo.subPath === `/`)
-      return discard();
-
-    return this.getZipSync(zipInfo.archivePath, zipFs => accept(zipFs, zipInfo));
-  }
-
-  private findZip(p: PortablePath) {
-    if (this.filter && !this.filter.test(p))
-      return null;
-
-    let filePath = `` as PortablePath;
-
-    while (true) {
-      const pathPartWithArchive = p.substring(filePath.length);
-
-      let archivePart;
-      if (!this.fileExtensions) {
-        archivePart = getArchivePart(pathPartWithArchive, `.zip`);
-      } else {
-        for (const ext of this.fileExtensions) {
-          archivePart = getArchivePart(pathPartWithArchive, ext);
-          if (archivePart) {
-            break;
+    const getMountPoint: GetMountPointFn = typeof fileExtensions === `undefined`
+      ? path => getArchivePart(path, `.zip`)
+      : path => {
+        for (const extension of fileExtensions!) {
+          const result = getArchivePart(path, extension);
+          if (result) {
+            return result;
           }
         }
-      }
 
-      if (!archivePart)
         return null;
-
-      filePath = this.pathUtils.join(filePath, archivePart);
-
-      if (this.isZip.has(filePath) === false) {
-        if (this.notZip.has(filePath))
-          continue;
-
-        try {
-          if (!this.baseFs.lstatSync(filePath).isFile()) {
-            this.notZip.add(filePath);
-            continue;
-          }
-        } catch {
-          return null;
-        }
-
-        this.isZip.add(filePath);
-      }
-
-      return {
-        archivePath: filePath,
-        subPath: this.pathUtils.join(PortablePath.root, p.substring(filePath.length) as PortablePath),
       };
-    }
-  }
 
-  private limitOpenFilesTimeout: NodeJS.Timeout | null = null;
-  private limitOpenFiles(max: number | null) {
-    if (this.zipInstances === null)
-      return;
+    const factorySync = (baseFs: FakeFS<PortablePath>, p: PortablePath) => {
+      return new ZipFS(p, {
+        baseFs,
+        libzip: getLibzip(),
+        readOnly: readOnlyArchives,
+        stats: baseFs.statSync(p),
+      });
+    };
 
-    const now = Date.now();
-    let nextExpiresAt = now + this.maxAge;
-    let closeCount = max === null ? 0 : this.zipInstances.size - max;
+    const factoryPromise = async (baseFs: FakeFS<PortablePath>, p: PortablePath) => {
+      const zipOptions = {
+        baseFs,
+        libzip: getLibzip(),
+        readOnly: readOnlyArchives,
+        stats: await baseFs.statPromise(p),
+      };
 
-    for (const [path, {zipFs, expiresAt, refCount}] of this.zipInstances.entries()) {
-      if (refCount !== 0 || zipFs.hasOpenFileHandles()) {
-        continue;
-      } else if (now >= expiresAt) {
-        zipFs.saveAndClose();
-        this.zipInstances.delete(path);
-        closeCount -= 1;
-        continue;
-      } else if (max === null || closeCount <= 0) {
-        nextExpiresAt = expiresAt;
-        break;
-      }
+      return () => {
+        return new ZipFS(p, zipOptions);
+      };
+    };
 
-      zipFs.saveAndClose();
-      this.zipInstances.delete(path);
-      closeCount -= 1;
-    }
+    super({
+      ...opts,
 
-    if (this.limitOpenFilesTimeout === null && ((max === null && this.zipInstances.size > 0) || max !== null)) {
-      this.limitOpenFilesTimeout = setTimeout(() => {
-        this.limitOpenFilesTimeout = null;
-        this.limitOpenFiles(null);
-      }, nextExpiresAt - now).unref();
-    }
-  }
+      factorySync,
+      factoryPromise,
 
-  private async getZipPromise<T>(p: PortablePath, accept: (zipFs: ZipFS) => Promise<T>) {
-    const getZipOptions = async () => ({
-      baseFs: this.baseFs,
-      libzip: this.libzip,
-      readOnly: this.readOnlyArchives,
-      stats: await this.baseFs.statPromise(p),
+      getMountPoint,
     });
-
-    if (this.zipInstances) {
-      let cachedZipFs = this.zipInstances.get(p);
-
-      if (!cachedZipFs) {
-        const zipOptions = await getZipOptions();
-
-        // We need to recheck because concurrent getZipPromise calls may
-        // have instantiated the zip archive while we were waiting
-        cachedZipFs = this.zipInstances.get(p);
-        if (!cachedZipFs) {
-          cachedZipFs = {
-            zipFs: new ZipFS(p, zipOptions),
-            expiresAt: 0,
-            refCount: 0,
-          };
-        }
-      }
-
-      // Removing then re-adding the field allows us to easily implement
-      // a basic LRU garbage collection strategy
-      this.zipInstances.delete(p);
-      this.limitOpenFiles(this.maxOpenFiles - 1);
-      this.zipInstances.set(p, cachedZipFs);
-
-      cachedZipFs.expiresAt = Date.now() + this.maxAge;
-      cachedZipFs.refCount += 1;
-      try {
-        return await accept(cachedZipFs.zipFs);
-      } finally {
-        cachedZipFs.refCount -= 1;
-      }
-    } else {
-      const zipFs = new ZipFS(p, await getZipOptions());
-
-      try {
-        return await accept(zipFs);
-      } finally {
-        zipFs.saveAndClose();
-      }
-    }
-  }
-
-  private getZipSync<T>(p: PortablePath, accept: (zipFs: ZipFS) => T) {
-    const getZipOptions = () => ({
-      baseFs: this.baseFs,
-      libzip: this.libzip,
-      readOnly: this.readOnlyArchives,
-      stats: this.baseFs.statSync(p),
-    });
-
-    if (this.zipInstances) {
-      let cachedZipFs = this.zipInstances.get(p);
-
-      if (!cachedZipFs) {
-        cachedZipFs = {
-          zipFs: new ZipFS(p, getZipOptions()),
-          expiresAt: 0,
-          refCount: 0,
-        };
-      }
-
-      // Removing then re-adding the field allows us to easily implement
-      // a basic LRU garbage collection strategy
-      this.zipInstances.delete(p);
-      this.limitOpenFiles(this.maxOpenFiles - 1);
-      this.zipInstances.set(p, cachedZipFs);
-
-      cachedZipFs.expiresAt = Date.now() + this.maxAge;
-      return accept(cachedZipFs.zipFs);
-    } else {
-      const zipFs = new ZipFS(p, getZipOptions());
-
-      try {
-        return accept(zipFs);
-      } finally {
-        zipFs.saveAndClose();
-      }
-    }
   }
 }
